UPLOAD to C:\\upload\  --- Read the description
In this commit we are now uploading files to the system. because we have
two computers the path of the upload file will not be the same. So I did
it using the root directory C

Create a folder called upload in the C drive then try to upload a
document and see if the file uploaded or not.